refactor: PathDrawingEngin을 interface에서 abstract로 변경
